requirements.txt
sqlalchemy
pyodbc
pandas
numpy
scikit-learn
python-dotenv
Faker
